% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{The Simplest and Most Correct Way to Do Causal Mediation Analysis}
\usage{
mediation(data = easymediation::fake_mediation_data)
}
\arguments{
\item{data}{Not needed, wouldn't make your analysis and more correct.}
}
\value{
Nothing, because you shouldn't do mediation analysis.
}
\description{
This function discourages you from performing mediation analysis
by providing warnings to you shouldn't do it. Easy, right?
}
\details{
We have:
\itemize{
\item X: a binary treatment variable, that is affected by an unmeasured confounder (U).
\item M: a mediator, that is mostly driven by U, not X
\item Y: an outcome, that is influenced by both M and U, but you can’t see U.
}

What you don’t have:
\itemize{
\item Any valid identification strategy
\item Temporal ordering
\item Careful causal reasoning
}
}
\examples{
mediation()
mediation(data = fake_mediation_data)
}
